Created social distancing helper class, removed not required changes
