+++ app/src/main/java/nic/goi/aarogyasetu/db/dao/WhiteListDataDao.java
@@ -10,16 +10,28 @@ import java.util.List;
 
 import nic.goi.aarogyasetu.models.WhiteListData;
 
+
+/**
+ * The interface WhiteListDataDao to access whiteList DB table.
+ */
 @Dao
 public interface WhiteListDataDao {
 
+    /**
+     * Insert white list device .
+     *
+     * @param device the device
+     * @return the long uniqueId of inserted device
+     */
     @Insert
     long insertWhiteListDevice(WhiteListData device);
 
+    /**
+     * Gets all white list devices.
+     *
+     * @return the all white list devices
+     */
     @Query("SELECT deviceId FROM white_list_devices")
     List<String> getAllWhiteListDevices();
-
-    @Delete
-    void deleteAll(List<WhiteListData> whiteListData);
 
 }
