+++ app/src/main/res/values-pu/strings.xml
@@ -117,7 +117,7 @@
     <string name="title_distance_breach">ਸਾਵਧਾਨ ਰਹੋ !!</string>
     <string name="multiple_breaches_message">
 ਤੁਸੀਂ ਹੋਰ ਆਰੋਗਿਆ ਸੇਤੂ ਉਪਭੋਗਤਾਵਾਂ ਦੇ ਨੇੜੇ ਹੋ ਰਹੇ ਹੋ, ਕਿਰਪਾ ਕਰਕੇ ਸੁਰੱਖਿਅਤ ਦੂਰੀ ਬਣਾਈ ਰੱਖੋ.
-ਜੇ ਉਹ ਤੁਹਾਡੇ ਪਰਿਵਾਰਕ ਮੈਂਬਰ ਹਨ, ਤਾਂ ਉਹਨਾਂ ਲੋਕਾਂ ਲਈ ਅੱਗੇ ਤੋਂ ਚਿਤਾਵਨੀਆਂ ਤੋਂ ਬਚਣ ਲਈ ਉਨ੍ਹਾਂ ਨੂੰ "ਪਰਿਵਾਰ" ਵਜੋਂ ਸ਼ਾਮਲ ਕਰੋ.</string>
+ਜੇ ਉਹ ਤੁਹਾਡੇ ਪਰਿਵਾਰਕ ਮੈਂਬਰ ਹਨ, ਤਾਂ ਉਹਨਾਂ ਲੋਕਾਂ ਲਈ ਅੱਗੇ ਤੋਂ ਚਿਤਾਵਨੀਆਂ ਤੋਂ ਬਚਣ ਲਈ ਉਨ੍ਹਾਂ ਨੂੰ \"ਪਰਿਵਾਰ\" ਵਜੋਂ ਸ਼ਾਮਲ ਕਰੋ.</string>
     <string name="single_breache_message">ਤੁਸੀਂ ਕਿਸੇ ਹੋਰ ਆਰੋਗਿਆ ਸੇਤੂ ਉਪਭੋਗਤਾ ਦੇ ਨੇੜੇ ਹੋ ਰਹੇ ਹੋ, ਕਿਰਪਾ ਕਰਕੇ ਸੁਰੱਖਿਅਤ ਦੂਰੀ ਬਣਾਈ ਰੱਖੋ.
-ਜੇ ਉਹ ਵਿਅਕਤੀ ਤੁਹਾਡਾ ਪਰਿਵਾਰਕ ਮੈਂਬਰ ਹੈ, ਤਾਂ ਇਸ ਵਿਅਕਤੀ ਲਈ ਵਧੇਰੇ ਚੇਤਾਵਨੀਆਂ ਤੋਂ ਬਚਣ ਲਈ ਉਸ ਵਿਅਕਤੀ ਨੂੰ '' ਪਰਿਵਾਰ '' ਵਜੋਂ ਸ਼ਾਮਲ ਕਰੋ.</string>
+ਜੇ ਉਹ ਵਿਅਕਤੀ ਤੁਹਾਡਾ ਪਰਿਵਾਰਕ ਮੈਂਬਰ ਹੈ, ਤਾਂ ਇਸ ਵਿਅਕਤੀ ਲਈ ਵਧੇਰੇ ਚੇਤਾਵਨੀਆਂ ਤੋਂ ਬਚਣ ਲਈ ਉਸ ਵਿਅਕਤੀ ਨੂੰ \"ਪਰਿਵਾਰ\" ਵਜੋਂ ਸ਼ਾਮਲ ਕਰੋ.</string>
 </resources>
